% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barnsley_fern.R
\docType{data}
\name{barnsley_fern}
\alias{barnsley_fern}
\title{Barnsley fern}
\format{object of class \code{IFS}}
\usage{
data(barnsley_fern)
}
\description{
An object of class \code{IFSflorek} holding functions to create Barnsley Fern.
}
\references{
https://en.wikipedia.org/wiki/Barnsley_fern
}
\keyword{datasets}
