% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIFS.R
\name{createIFS}
\alias{createIFS}
\title{Creates class "IFS" from list of functions.}
\usage{
createIFS(f1, ...)
}
\arguments{
\item{f1}{function (contraction) used to create an IFS}

\item{...}{possibly more functions (contractions)}
}
\value{
An object of class \code{IFS} with a list of functions.
}
\description{
Creates class "IFS" from list of functions.
}
