% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIFS_from_coefficients.R
\name{createIFS_from_coefficients}
\alias{createIFS_from_coefficients}
\title{Creates class "IFS" from a matrix of coefficients.}
\usage{
createIFS_from_coefficients(params)
}
\arguments{
\item{params}{matrix of parameters for contraction mappings}
}
\value{
An object of class \code{IFS} with a list of functions.
}
\description{
Creates class "IFS" from a matrix of coefficients.
}
