% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleIFS.R
\name{exampleIFS}
\alias{exampleIFS}
\title{This function can draw some fractals.}
\usage{
exampleIFS(fractal_name, depth = 8)
}
\arguments{
\item{fractal_name}{Name of a fractal}

\item{depth}{depth to which the fractal should be drawn. Defaults to 8.}
}
\description{
This function can draw some fractals.
}
\details{
Currently available are following fractals:
\itemize{
  \item sierpinski_gasket (or sierpinski_triangle, or sierpinski)
  \item fractal_tree
  \item koch (or koch_curve)
  \item heighway_dragon
  \item barnsley_fern
}
}
