% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal_tree.R
\docType{data}
\name{fractal_tree}
\alias{fractal_tree}
\title{Tree}
\format{object of class \code{IFSflorek}}
\usage{
data(fractal_tree)
}
\description{
An object of class \code{IFSflorek} holding functions to create a fractal tree.
}
\keyword{datasets}
