% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heighway_dragon.R
\docType{data}
\name{heighway_dragon}
\alias{heighway_dragon}
\title{Heighway Dragon}
\format{object of class \code{IFS}}
\usage{
data(heighway_dragon)
}
\description{
An object of class \code{IFSflorek} holding functions to create Heighway Dragon,
also known as the Harter–Heighway dragon or the Jurassic Park dragon.
}
\references{
https://en.wikipedia.org/wiki/Dragon_curve#Heighway_dragon
}
\keyword{datasets}
