% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koch.R
\docType{data}
\name{koch}
\alias{koch}
\title{Koch curve}
\format{object of class \code{IFS}}
\usage{
data(koch)
}
\description{
An object of class \code{IFSflorek} holding functions to create Koch curve.
Not to be mistaken with Koch snowflake.
}
\references{
https://en.wikipedia.org/wiki/Koch_snowflake
}
\keyword{datasets}
