% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIFS.R
\name{plot.IFS}
\alias{plot.IFS}
\title{Plot a fractal from IFS Object}
\usage{
\method{plot}{IFS}(x, depth = 8, starting_point = c(0, 0),
  colour_palette = rainbow, reverse = TRUE, ...)
}
\arguments{
\item{x}{object of a class \code{IFS}}

\item{depth}{depth to which the fractal should be plotted. Defaults to 8.}

\item{starting_point}{starting point to which the functions of \code{x} should be applied}

\item{colour_palette}{function defining colours to be used for each iteration or a list of colours; defaults to rainbow}

\item{reverse}{wheter the iterations should be plotted and coloured in reversed order; default: TRUE}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
plot.IFS is an overloaded plotting function for \code{IFS} objects.
}
\examples{
fractal <- createIFS( function(x) { c(0.5*x[1], 0.5*x[2]) },
function(x) { c(0.5*x[1] + 0.5, 0.5*x[2]) },
function(x) { c(0.5*x[1] + 0.25, 0.5*x[2] + sqrt(3)/4) } )
plot(fractal, depth = 9)

}
