% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierpinski_gasket.R
\docType{data}
\name{sierpinski_gasket}
\alias{sierpinski_gasket}
\title{Sierpinski's Gasket}
\format{object of class \code{IFSflorek}}
\usage{
data(sierpinski_gasket)
}
\description{
An object of class \code{IFSflorek} holding functions to create Sierpinski's Gasket.
}
\references{
https://en.wikipedia.org/wiki/Sierpinski_triangle
}
\keyword{datasets}
