% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transformation}
\alias{transformation}
\title{Linear transformation}
\usage{
transformation(x, a, b, c, d, e = 0, f = 0)
}
\arguments{
\item{x}{two-element vector to transform}

\item{a, b, c, d, e, f}{coefficients of the transformation}
}
\description{
Transforms two-element vector with a linear function of a form:
f(x, y) = (ax + by + e, cx + dy + f)
}
